requirements.txt
# Cardigan v4.0 - Python Dependencies

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# HTTP Client (for LLM APIs)
httpx>=0.26.0
requests>=2.31.0

# HTML Parsing (for ingest server directory listing)
beautifulsoup4>=4.12.0

# Scheduling (for ingest scanner)
apscheduler>=3.10.0

# WebSocket support
websockets>=12.0

# Rate limiting
slowapi>=0.1.9

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Linting/Formatting
ruff>=0.1.0
black>=23.12.0

# Development
watchfiles>=0.21.0

# MCP Server (for Claude Desktop integration)
mcp>=1.0.0
